+++ visionforge-core/src/commonMain/kotlin/hep/dataforge/vision/VisionGroup.kt
@@ -3,7 +3,7 @@ package hep.dataforge.vision
 import hep.dataforge.names.*
 import hep.dataforge.provider.Provider
 
-public interface VisionContainer<out V: Vision>{
+public interface VisionContainer<out V : Vision> {
     public operator fun get(name: Name): V?
 }
 
@@ -19,12 +19,6 @@ public interface VisionGroup : Provider, Vision, VisionContainer<Vision> {
     override val defaultTarget: String get() = Vision.TYPE
 
     /**
-     * A stylesheet for this group and its descendants. Stylesheet is not applied directly,
-     * but instead is just a repository for named configurations.
-     */
-    public val styleSheet: StyleSheet?
-
-    /**
      * A map of direct children for specific target
      * (currently "visual" or "style")
      */
@@ -38,7 +32,7 @@ public interface VisionGroup : Provider, Vision, VisionContainer<Vision> {
                 }
                 res.entries
             }.associate { it.toPair() }
-            STYLE_TARGET -> styleSheet?.items?.mapKeys { it.key.toName() } ?: emptyMap()
+            STYLE_TARGET -> styleSheet.items?.mapKeys { it.key.asName() } ?: emptyMap()
             else -> emptyMap()
         }
 
@@ -54,7 +48,6 @@ public interface VisionGroup : Provider, Vision, VisionContainer<Vision> {
      * A fix for serialization bug that writes all proper parents inside the tree after deserialization
      */
     public fun attachChildren() {
-        styleSheet?.owner = this
         children.values.forEach {
             it.parent = this
             (it as? VisionGroup)?.attachChildren()
@@ -73,7 +66,7 @@ public operator fun VisionGroup.iterator(): Iterator<Vision> = children.values.i
 
 public val VisionGroup.isEmpty: Boolean get() = this.children.isEmpty()
 
-public interface VisionContainerBuilder<in V: Vision>{
+public interface VisionContainerBuilder<in V : Vision> {
     public operator fun set(name: Name, child: V?)
 }
 
@@ -97,9 +90,11 @@ public interface MutableVisionGroup : VisionGroup, VisionContainerBuilder<Vision
 //    public operator fun set(name: Name, child: Vision?)
 }
 
-public operator fun <V: Vision> VisionContainer<V>.get(str: String): V? = get(str.toName())
+public operator fun <V : Vision> VisionContainer<V>.get(str: String): V? = get(str.toName())
 
-public operator fun <V: Vision>  VisionContainerBuilder<V>.set(token: NameToken, child: V?): Unit = set(token.asName(), child)
-public operator fun <V: Vision>  VisionContainerBuilder<V>.set(key: String, child: V?): Unit = set(key.toName(), child)
+public operator fun <V : Vision> VisionContainerBuilder<V>.set(token: NameToken, child: V?): Unit =
+    set(token.asName(), child)
+
+public operator fun <V : Vision> VisionContainerBuilder<V>.set(key: String, child: V?): Unit = set(key.toName(), child)
 
 public fun MutableVisionGroup.removeAll(): Unit = children.keys.map { it.asName() }.forEach { this[it] = null }